requirements.txt
python-dotenv==0.17.0
kaggle==1.5.12
scikit-learn==0.24.1
pandas==1.2.1
numpy==1.18.5